wellness/requirements.txt
fastapi
uvicorn
pydantic
google-adk
python-dotenv
requests
nest-asyncio
